switch everything to float32

+++ internal/camera/camera.go
@@ -11,9 +11,9 @@ import (
 
 // Camera represents a camera in the world.
 type Camera struct {
-	lensRadius      float64
-	time0           float64
-	time1           float64
+	lensRadius      float32
+	time0           float32
+	time1           float32
 	u               *vec3.Vec3Impl
 	v               *vec3.Vec3Impl
 	origin          *vec3.Vec3Impl
@@ -24,11 +24,11 @@ type Camera struct {
 
 // New returns an instance of a camera.
 func New(lookFrom *vec3.Vec3Impl, lookAt *vec3.Vec3Impl, vup *vec3.Vec3Impl,
-	vfov float64, aspect float64, aperture float64, focusDist float64, time0 float64, time1 float64) *Camera {
+	vfov float32, aspect float32, aperture float32, focusDist float32, time0 float32, time1 float32) *Camera {
 
 	lensRadius := aperture / 2.0
-	theta := vfov * math.Pi / 180
-	halfHeight := math.Tan(theta / 2.0)
+	theta := vfov * float32(math.Pi) / 180
+	halfHeight := float32(math.Tan(float64(theta) / 2.0))
 	halfWidth := aspect * halfHeight
 	w := vec3.UnitVector(vec3.Sub(lookFrom, lookAt))
 	u := vec3.UnitVector(vec3.Cross(vup, w))
@@ -54,10 +54,10 @@ func New(lookFrom *vec3.Vec3Impl, lookAt *vec3.Vec3Impl, vup *vec3.Vec3Impl,
 }
 
 // GetRay returns the ray associated for the supplied u and v.
-func (c *Camera) GetRay(s float64, t float64) *ray.RayImpl {
+func (c *Camera) GetRay(s float32, t float32) *ray.RayImpl {
 	rd := vec3.ScalarMul(randomInUnitDisc(), c.lensRadius)
 	offset := vec3.Add(vec3.ScalarMul(c.u, rd.X), vec3.ScalarMul(c.v, rd.Y))
-	time := c.time0 + rand.Float64()*(c.time1-c.time0)
+	time := c.time0 + rand.Float32()*(c.time1-c.time0)
 	return ray.New(vec3.Add(c.origin, offset),
 		// lowerLeftCorner + s*horizontal + t*vertical - origin - offset
 		vec3.Sub(vec3.Add(c.lowerLeftCorner, vec3.ScalarMul(c.horizontal, s),
@@ -65,10 +65,10 @@ func (c *Camera) GetRay(s float64, t float64) *ray.RayImpl {
 }
 
 // GetRayWithLambda returns the ray associated for the supplied u and v with a specific wavelength.
-func (c *Camera) GetRayWithLambda(s float64, t float64, lambda float64) *ray.RayImpl {
+func (c *Camera) GetRayWithLambda(s float32, t float32, lambda float32) *ray.RayImpl {
 	rd := vec3.ScalarMul(randomInUnitDisc(), c.lensRadius)
 	offset := vec3.Add(vec3.ScalarMul(c.u, rd.X), vec3.ScalarMul(c.v, rd.Y))
-	time := c.time0 + rand.Float64()*(c.time1-c.time0)
+	time := c.time0 + rand.Float32()*(c.time1-c.time0)
 	return ray.NewWithLambda(vec3.Add(c.origin, offset),
 		// lowerLeftCorner + s*horizontal + t*vertical - origin - offset
 		vec3.Sub(vec3.Add(c.lowerLeftCorner, vec3.ScalarMul(c.horizontal, s),
@@ -77,7 +77,7 @@ func (c *Camera) GetRayWithLambda(s float64, t float64, lambda float64) *ray.Ray
 
 func randomInUnitDisc() *vec3.Vec3Impl {
 	for {
-		p := vec3.Sub(vec3.ScalarMul(&vec3.Vec3Impl{X: rand.Float64(), Y: rand.Float64()}, 2.0), &vec3.Vec3Impl{X: 1.0, Y: 1.0})
+		p := vec3.Sub(vec3.ScalarMul(&vec3.Vec3Impl{X: rand.Float32(), Y: rand.Float32()}, 2.0), &vec3.Vec3Impl{X: 1.0, Y: 1.0})
 		if vec3.Dot(p, p) < 1.0 {
 			return p
 		}
