+++ internal/hitable/box.go
@@ -41,19 +41,19 @@ func NewBox(p0 *vec3.Vec3Impl, p1 *vec3.Vec3Impl, mat material.Material) *Box {
 	}
 }
 
-func (b *Box) Hit(r ray.Ray, tMin float64, tMax float64) (*hitrecord.HitRecord, material.Material, bool) {
+func (b *Box) Hit(r ray.Ray, tMin float32, tMax float32) (*hitrecord.HitRecord, material.Material, bool) {
 	return b.sides.Hit(r, tMin, tMax)
 }
 
-func (b *Box) HitEdge(r ray.Ray, tMin float64, tMax float64) (*hitrecord.HitRecord, bool, bool) {
+func (b *Box) HitEdge(r ray.Ray, tMin float32, tMax float32) (*hitrecord.HitRecord, bool, bool) {
 	return b.sides.HitEdge(r, tMin, tMax)
 }
 
-func (b *Box) BoundingBox(time0 float64, time1 float64) (*aabb.AABB, bool) {
+func (b *Box) BoundingBox(time0 float32, time1 float32) (*aabb.AABB, bool) {
 	return b.sides.BoundingBox(time0, time1)
 }
 
-func (b *Box) PDFValue(o *vec3.Vec3Impl, v *vec3.Vec3Impl) float64 {
+func (b *Box) PDFValue(o *vec3.Vec3Impl, v *vec3.Vec3Impl) float32 {
 	return 0.0
 }
