+++ internal/hitable/bvh_node.go
@@ -23,21 +23,21 @@ var _ Hitable = (*BVHNode)(nil)
 type BVHNode struct {
 	left  Hitable
 	right Hitable
-	time0 float64
-	time1 float64
+	time0 float32
+	time1 float32
 	box   *aabb.AABB
 }
 
-func NewBVH(hitables []Hitable, time0 float64, time1 float64) *BVHNode {
+func NewBVH(hitables []Hitable, time0 float32, time1 float32) *BVHNode {
 	log.Infof("Building BVH with %v elements", len(hitables))
 	startTime := time.Now()
-	randomFunc := rand.Float64
+	randomFunc := rand.Float32
 	bvh := newBVH(hitables, randomFunc, time0, time1)
 	log.Infof("Completed BVH construction in %v", time.Since(startTime))
 	return bvh
 }
 
-func newBVH(hitables []Hitable, randomFunc func() float64, time0 float64, time1 float64) *BVHNode {
+func newBVH(hitables []Hitable, randomFunc func() float32, time0 float32, time1 float32) *BVHNode {
 	bn := &BVHNode{
 		time0: time0,
 		time1: time1,
@@ -130,7 +130,7 @@ func newBVH(hitables []Hitable, randomFunc func() float64, time0 float64, time1 
 	return bn
 }
 
-func (bn *BVHNode) Hit(r ray.Ray, tMin float64, tMax float64) (*hitrecord.HitRecord, material.Material, bool) {
+func (bn *BVHNode) Hit(r ray.Ray, tMin float32, tMax float32) (*hitrecord.HitRecord, material.Material, bool) {
 	if bn.box.Hit(r, tMin, tMax) {
 		leftRec, leftMat, hitLeft := bn.left.Hit(r, tMin, tMax)
 		rightRec, rightMat, hitRight := bn.right.Hit(r, tMin, tMax)
@@ -154,7 +154,7 @@ func (bn *BVHNode) Hit(r ray.Ray, tMin float64, tMax float64) (*hitrecord.HitRec
 	return nil, nil, false
 }
 
-func (bn *BVHNode) HitEdge(r ray.Ray, tMin float64, tMax float64) (*hitrecord.HitRecord, bool, bool) {
+func (bn *BVHNode) HitEdge(r ray.Ray, tMin float32, tMax float32) (*hitrecord.HitRecord, bool, bool) {
 	if bn.box.Hit(r, tMin, tMax) {
 		leftRec, hitLeft, hitLeftEdge := bn.left.HitEdge(r, tMin, tMax)
 		rightRec, hitRight, hitRightEdge := bn.right.HitEdge(r, tMin, tMax)
@@ -178,11 +178,11 @@ func (bn *BVHNode) HitEdge(r ray.Ray, tMin float64, tMax float64) (*hitrecord.Hi
 	return nil, false, false
 }
 
-func (bn *BVHNode) BoundingBox(time0 float64, time1 float64) (*aabb.AABB, bool) {
+func (bn *BVHNode) BoundingBox(time0 float32, time1 float32) (*aabb.AABB, bool) {
 	return bn.box, true
 }
 
-func (bn *BVHNode) PDFValue(o *vec3.Vec3Impl, v *vec3.Vec3Impl) float64 {
+func (bn *BVHNode) PDFValue(o *vec3.Vec3Impl, v *vec3.Vec3Impl) float32 {
 	return 0.0
 }
