+++ internal/material/diffuselight.go
@@ -46,7 +46,7 @@ func (dl *DiffuseLight) SpectralScatter(_ ray.Ray, _ *hitrecord.HitRecord, _ *fa
 }
 
 // Emitted returns the texture value at that point.
-func (dl *DiffuseLight) Emitted(rIn ray.Ray, rec *hitrecord.HitRecord, u float64, v float64, p *vec3.Vec3Impl) *vec3.Vec3Impl {
+func (dl *DiffuseLight) Emitted(rIn ray.Ray, rec *hitrecord.HitRecord, u float32, v float32, p *vec3.Vec3Impl) *vec3.Vec3Impl {
 	if vec3.Dot(rec.Normal(), rIn.Direction()) < 0.0 {
 		return dl.emit.Value(u, v, p)
 	}
@@ -55,7 +55,7 @@ func (dl *DiffuseLight) Emitted(rIn ray.Ray, rec *hitrecord.HitRecord, u float64
 }
 
 // EmittedSpectral returns the spectral emission at the given wavelength for diffuse lights.
-func (dl *DiffuseLight) EmittedSpectral(rIn ray.Ray, rec *hitrecord.HitRecord, u float64, v float64, lambda float64, p *vec3.Vec3Impl) float64 {
+func (dl *DiffuseLight) EmittedSpectral(rIn ray.Ray, rec *hitrecord.HitRecord, u float32, v float32, lambda float32, p *vec3.Vec3Impl) float32 {
 	if vec3.Dot(rec.Normal(), rIn.Direction()) < 0.0 {
 		return dl.spectralEmit.Value(u, v, lambda, p)
 	}
@@ -63,7 +63,7 @@ func (dl *DiffuseLight) EmittedSpectral(rIn ray.Ray, rec *hitrecord.HitRecord, u
 }
 
 // ScatteringPDF implements the probability distribution function for diffuse lights.
-func (dl *DiffuseLight) ScatteringPDF(r ray.Ray, hr *hitrecord.HitRecord, scattered ray.Ray) float64 {
+func (dl *DiffuseLight) ScatteringPDF(r ray.Ray, hr *hitrecord.HitRecord, scattered ray.Ray) float32 {
 	return 0
 }
 
@@ -71,11 +71,11 @@ func (dl *DiffuseLight) IsEmitter() bool {
 	return true
 }
 
-func (dl *DiffuseLight) Albedo(u float64, v float64, p *vec3.Vec3Impl) *vec3.Vec3Impl {
+func (dl *DiffuseLight) Albedo(u float32, v float32, p *vec3.Vec3Impl) *vec3.Vec3Impl {
 	return dl.emit.Value(u, v, p)
 }
 
 // SpectralAlbedo returns the spectral albedo at the given wavelength.
-func (dl *DiffuseLight) SpectralAlbedo(u float64, v float64, lambda float64, p *vec3.Vec3Impl) float64 {
+func (dl *DiffuseLight) SpectralAlbedo(u float32, v float32, lambda float32, p *vec3.Vec3Impl) float32 {
 	return dl.emit.Value(u, v, p).X // Use red component as approximation
 }
