+++ internal/material/metal.go
@@ -19,11 +19,11 @@ type Metal struct {
 	nonPathLength
 	nonWorldSetter
 	albedo *vec3.Vec3Impl
-	fuzz   float64
+	fuzz   float32
 }
 
 // NewMetal returns an instance of the metal material.
-func NewMetal(albedo *vec3.Vec3Impl, fuzz float64) *Metal {
+func NewMetal(albedo *vec3.Vec3Impl, fuzz float32) *Metal {
 	return &Metal{
 		albedo: albedo,
 		fuzz:   fuzz,
@@ -40,16 +40,16 @@ func (m *Metal) Scatter(r ray.Ray, hr *hitrecord.HitRecord, random *fastrandom.L
 }
 
 // ScatteringPDF implements the probability distribution function for metals.
-func (m *Metal) ScatteringPDF(r ray.Ray, hr *hitrecord.HitRecord, scattered ray.Ray) float64 {
+func (m *Metal) ScatteringPDF(r ray.Ray, hr *hitrecord.HitRecord, scattered ray.Ray) float32 {
 	return 0
 }
 
-func (m *Metal) Albedo(u float64, v float64, p *vec3.Vec3Impl) *vec3.Vec3Impl {
+func (m *Metal) Albedo(u float32, v float32, p *vec3.Vec3Impl) *vec3.Vec3Impl {
 	a := *m.albedo
 	return &a
 }
 
 // SpectralAlbedo returns the spectral albedo at the given wavelength.
-func (m *Metal) SpectralAlbedo(u float64, v float64, lambda float64, p *vec3.Vec3Impl) float64 {
+func (m *Metal) SpectralAlbedo(u float32, v float32, lambda float32, p *vec3.Vec3Impl) float32 {
 	return m.albedo.X // Use red component as approximation
 }
