+++ internal/pdf/api.go
@@ -9,7 +9,7 @@ import (
 // PDF represents a probability density function.
 type PDF interface {
 	// Value computes the probability density function at a given point.
-	Value(direction *vec3.Vec3Impl) float64
+	Value(direction *vec3.Vec3Impl) float32
 	// Generate generates a probability density function.
 	Generate(random *fastrandom.LCG) *vec3.Vec3Impl
 }
